Refactor index.html with js modules in index.js
