+++ modules/getBooksStore.js
@@ -1,0 +1,4 @@
+const awesomeBooksIn = new BooksClass();
+if (awesomeBooksIn.getLocalStorage().length > 0) {
+  awesomeBooksIn.ShowAllBooks();
+}
