+++ modules/getCurrentDate.js
@@ -1,0 +1,5 @@
+// get date from luxon
+const date = new Date();
+currentDate.append(document.createElement('p').innerHTML = `${date}`);
+
+<div class="current-date"></div>
